fix: change product to have at least one option

+++ src/main/java/gift/service/OptionService.java
@@ -3,6 +3,7 @@ package gift.service;
 import gift.domain.Option;
 import gift.domain.Product;
 import gift.dto.OptionDTO;
+import gift.exception.NoOptionsForProductException;
 import gift.exception.NoSuchOptionException;
 import gift.exception.NoSuchProductException;
 import gift.repository.OptionRepository;
@@ -45,7 +46,12 @@ public class OptionService {
         return optionRepository.save(updatedOption).toDTO();
     }
 
-    public OptionDTO deleteOption(long id) {
+    public OptionDTO deleteOption(long productId, long id) {
+        Product product = productRepository.findById(productId)
+            .orElseThrow(NoSuchProductException::new);
+        if (product.getOptions().size() == 1) {
+            throw new NoOptionsForProductException();
+        }
         Option deletedOption = optionRepository.findById(id)
             .orElseThrow(NoSuchOptionException::new);
         optionRepository.delete(deletedOption);
